refactor: split checkpoint probe helpers
